# Vectorize eq_ignore_ascii_case