Process in chunks of 16 for vectorization. ~12x speedup on larger chunks. ~4x speedup when N = 16.

Benchmark on x86 on equal slices:

NEW:
```
test bench_eq_ignore_ascii_case_std_n_8192 ... bench:           436.24 ns/iter (+/- 14.15)
test bench_eq_ignore_ascii_case_std_n_1000 ... bench:           63.94 ns/iter (+/- 3.45)
test bench_eq_ignore_ascii_case_std_n_16 ... bench:             2.41 ns/iter (+/- 0.07)
test bench_eq_ignore_ascii_case_std_n_15 ... bench:             14.74 ns/iter (+/- 0.18)
test bench_eq_ignore_ascii_case_std_n_8  ... bench:             8.06 ns/iter (+/- 0.34)
test bench_eq_ignore_ascii_case_std_n_4  ... bench:             4.57 ns/iter (+/- 0.13)
```

OLD:
```
test bench_eq_ignore_ascii_case_std_n_8192 ... bench:           5,333.35 ns/iter (+/- 22.50)
test bench_eq_ignore_ascii_case_std_n_1000 ... bench:           692.75 ns/iter (+/- 46.33)
test bench_eq_ignore_ascii_case_std_n_16 ... bench:             10.38 ns/iter (+/- 0.07)
test bench_eq_ignore_ascii_case_std_n_15 ... bench:             9.61 ns/iter (+/- 0.21)
test bench_eq_ignore_ascii_case_std_n_8  ... bench:             5.23 ns/iter (+/- 0.17)
test bench_eq_ignore_ascii_case_std_n_4  ... bench:             3.05 ns/iter (+/- 0.03)
```

(When built with target-cpu=native, I get up to 24x speedups)

Notice the performance is slightly worse on small N. Seems quite large for what should be one extra branch.

Godbolt: https://rust.godbolt.org/z/sz3baEYxf