@tgross35 2025-03-17
The gains here look pretty great. However, it looks like codegen gets worse on platforms that don't have vector instructions https://rust.godbolt.org/z/T7qqqTca8. Is there a better way to hint to LLVM that vectorization should be possible without encouraging it quite as much?

@scottmcm do you have any ideas?