@LaihoE 2025-03-18
> That's been changing lately https://discourse.llvm.org/t/rfc-supporting-more-early-exit-loops/84690?u=scottmcm, but I don't know how much is there for us in LLVM20, or how to communicate the "yes, we promise this is initialized"-ness of the whole slice.

Oh this is very interesting, thanks for sharing. Didn't know that support for early exit loops had made this much progress lately 🤔

In that case I think this PR goes the wrong direction. Better to let LLVM do this if the support for it is being worked on.